+++ ai_spider/files.py
@@ -1,18 +1,19 @@
+import logging
 import os
 from functools import wraps
 
-import httpx
 from botocore.exceptions import ClientError
 from dotenv import load_dotenv
-from fastapi import File, UploadFile, HTTPException, Depends, Request, APIRouter
+from fastapi import File, UploadFile, HTTPException, Depends, APIRouter
 from fastapi.responses import StreamingResponse
 import boto3
 
-from ai_spider.util import get_bill_to, BILLING_URL, BILLING_TIMEOUT
+from ai_spider.util import check_bearer_token, USER_BUCKET_NAME
 
 app = APIRouter()
 
 load_dotenv()
+log = logging.getLogger(__name__)
 
 g_s3 = None
 
@@ -25,9 +26,6 @@ def s3():
                             aws_secret_access_key=os.environ.get("AWS_SECRET_ACCESS_KEY"))
 
     return g_s3
-
-
-bucket_name = os.environ.get("GPUTOPIA_USER_BUCKET", 'gputopia-user-bucket')
 
 
 def handle_aws_exceptions(route_function):
@@ -47,35 +45,11 @@ def handle_aws_exceptions(route_function):
     return wrapper
 
 
-async def check_bearer_token(request: Request) -> str:
-    bill_to_token = get_bill_to(request)
-
-    if bill_to_token == os.environ.get("BYPASS_TOKEN"):
-        return "bypass"
-
-    command = dict(
-        command="check",
-        bill_to_token=bill_to_token,
-    )
-
-    try:
-        res = httpx.post(BILLING_URL, json=command, timeout=BILLING_TIMEOUT)
-    except Exception as ex:
-        raise HTTPException(status_code=500, detail="billing endpoint error: %s" % ex)
-
-    js = res.json()
-
-    if js.get("user_id"):
-        return js.get("user_id")
-
-    raise HTTPException(status_code=401, detail="Invalid token")
-
-
 @app.get("/files")
 @handle_aws_exceptions
 async def list_files(user_id: str = Depends(check_bearer_token)):
     user_folder = f"{user_id}/"
-    file_objects = s3().list_objects(Bucket=bucket_name, Prefix=user_folder)['Contents']
+    file_objects = s3().list_objects(Bucket=USER_BUCKET_NAME, Prefix=user_folder)['Contents']
     return {"data": [
         {"id": obj["Key"][len(user_folder):], "bytes": obj["Size"], "created_at": obj["LastModified"].timestamp()} for
         obj in
@@ -85,10 +59,12 @@ async def list_files(user_id: str = Depends(check_bearer_token)):
 @app.post("/files")
 @handle_aws_exceptions
 async def upload_file(file: UploadFile = File(...), purpose: str = "", user_id: str = Depends(check_bearer_token)):
+    # for now, this is a quick way of dealing with identical uploads
+    file_name = f"file_{file.size}"
     user_folder = f"{user_id}/"
-    s3().upload_fileobj(file.file, bucket_name, f"{user_folder}{file.filename}")
+    s3().upload_fileobj(file.file, USER_BUCKET_NAME, f"{user_folder}{file_name}")
     # You may also need to store additional metadata, such as purpose, in a database
-    return {"id": file.filename, "object": "file", "bytes": file.size, "filename": file.filename, "purpose": purpose,
+    return {"id": file_name, "object": "file", "bytes": file.size, "filename": file_name, "purpose": purpose,
             "status": "uploaded"}
 
 
@@ -97,7 +73,7 @@ async def upload_file(file: UploadFile = File(...), purpose: str = "", user_id: 
 async def delete_file(file_id: str, user_id: str = Depends(check_bearer_token)):
     try:
         user_folder = f"{user_id}/"
-        s3().delete_object(Bucket=bucket_name, Key=f"{user_folder}{file_id}")
+        s3().delete_object(Bucket=USER_BUCKET_NAME, Key=f"{user_folder}{file_id}")
         return {"id": file_id, "object": "file", "deleted": True}
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
@@ -109,7 +85,7 @@ async def retrieve_file(file_id: str, user_id: str = Depends(check_bearer_token)
     try:
         user_folder = f"{user_id}/"
         # Retrieve file metadata from S3 and potentially other metadata from the database
-        file_meta = s3().head_object(Bucket=bucket_name, Key=f"{user_folder}{file_id}")
+        file_meta = s3().head_object(Bucket=USER_BUCKET_NAME, Key=f"{user_folder}{file_id}")
         return {"id": file_id, "object": "file", "bytes": file_meta["ContentLength"],
                 "created_at": file_meta["LastModified"].timestamp(), "filename": file_id}
     except Exception as e:
@@ -120,5 +96,5 @@ async def retrieve_file(file_id: str, user_id: str = Depends(check_bearer_token)
 @handle_aws_exceptions
 async def retrieve_file_content(file_id: str, user_id: str = Depends(check_bearer_token)):
     user_folder = f"{user_id}/"
-    file_content = s3().get_object(Bucket=bucket_name, Key=f"{user_folder}{file_id}")['Body']
+    file_content = s3().get_object(Bucket=USER_BUCKET_NAME, Key=f"{user_folder}{file_id}")['Body']
     return StreamingResponse(content=file_content, media_type="application/octet-stream")
